Added JWT for authenticating user Modified contact schema to include contacts created by users Also made every CRUD method private i.e. ensured it goes through validate token

+++ controller/contactController.js
@@ -12,14 +12,15 @@ const VALUE_NOT_FOUND = "Contact not found.";
  * @access public
 */
 const getContacts = asyncHandler(async (req, res) => {
-    const contact = await Contact.find();
-    res.status(200).json(contact);
+    const contacts = await Contact.find({created_by: req.user.id});
+    res.status(200).json(contacts);
 })
 
 /**
  * @desc Get contact
  * @route GET /api/contacts/:id
- * @access public
+ * @access private
+ * @todo show contacts that only have been created by user
  */
 const getContact = asyncHandler(async (req, res) => {
     if (!(req.params.id.length == 24)) {
@@ -40,17 +41,23 @@ const getContact = asyncHandler(async (req, res) => {
 /**
  * @desc Create new contact
  * @route POST /api/contacts
- * @access public
+ * @access private
  */
 const createContact = asyncHandler(async (req, res) => {
     
     const { name, email, phone} = req.body;
+
     if (!name || !email || !phone) {
         res.status(400);
         throw new Error("All fields are mandatory.");
     }
-    
-    const contact = await Contact.create({name, email, phone});
+
+    const contact = await Contact.create({
+        name, 
+        email, 
+        phone,
+        created_by: req.user.id
+    });
     res.status(200).json(contact);
     
 })
@@ -58,13 +65,13 @@ const createContact = asyncHandler(async (req, res) => {
 /**
  * @desc Update contact
  * @route PUT /api/contacts/:id
- * @access public
+ * @access private
  */
 const updateContact = asyncHandler(async (req, res) => {
-    if (req.params.id.length != 25) {
-        res.status(400);
-        throw new Error(INVALID_ID);
-    }
+    // if (req.params.id.length != 25) {
+    //     res.status(400);
+    //     throw new Error(INVALID_ID);
+    // }
     
     const contact = await Contact.findById(req.params.id);
 
@@ -72,6 +79,11 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error(VALUE_NOT_FOUND)
     } 
+    
+    if (contact.created_by.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error("User is not allowed to update contacts created by others.");
+    }
     
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -85,7 +97,7 @@ const updateContact = asyncHandler(async (req, res) => {
 /**
  * @desc Update contact
  * @route DELETE /api/contacts/:id
- * @access public
+ * @access private
 */
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
@@ -93,6 +105,11 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error(VALUE_NOT_FOUND);
+    }
+
+    if (contact.created_by.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error("User is not allowed to delete contacts created by others.");
     }
 
     await Contact.findByIdAndDelete(req.params.id);
